% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_correction_matrix.R
\name{bias_correction_matrix}
\alias{bias_correction_matrix}
\title{Evaluate the bias correction matrix for the pseudocensus OLS estimator}
\usage{
bias_correction_matrix(x, recipients, donors, weights)
}
\arguments{
\item{x}{A numeric matrix of rowwise covariates used as matching variables
for distance calculations.}

\item{recipients}{An integer vector of row indices in \code{x} identifying the
reference (query) points.}

\item{donors}{An integer vector of row indices in \code{x} identifying the
candidate points.}

\item{weights}{A numeric matrix where \code{nrow(weights) = length(recipients)}
and \code{ncol(weights) <= length(donors)}. Each row in \code{weights} supplies the
imputation weights used in weighted nearest neighbours computation.
\code{weights} corresponds rowwise to \code{recipients} and columnwise to the rank of
the proximally sorted donors.}
}
\value{
A square numeric matrix with \code{ncol(x)} rows and columns.
}
\description{
Evaluate the bias correction matrix for the pseudocensus OLS estimator
}
