% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_nearer_neighbours.R
\name{count_nearer_neighbours}
\alias{count_nearer_neighbours}
\title{Count the number of points closer to the recipient}
\usage{
count_nearer_neighbours(donor, recipient, covariates, p = 2)
}
\arguments{
\item{donor}{The index of the prospective donor}

\item{recipient}{The index of the recipient}

\item{covariates}{A design matrix of covariates listed row-wise}

\item{p}{The order of the \eqn{\ell_{p}}-norm. The default value \code{p = 2}
yields the Euclidean metric.}
}
\value{
A non-negative integer.``li
}
\description{
\code{count_nearer_neighbours()} counts the number of points whose covariates are
relatively closer to the \code{recipient} covariate than the \code{donor} covariate
is to the  \code{recipient} covariate.
}
