% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{derive_global_debiasing_weights}
\alias{derive_global_debiasing_weights}
\title{Derive a global vector of weights that minimises aggregate covariate
imputation error}
\usage{
derive_global_debiasing_weights(
  x,
  recipients,
  donors,
  n_weights = length(donors)
)
}
\arguments{
\item{x}{A numeric matrix of rowwise covariates used as matching variables
for distance calculations.}

\item{recipients}{An integer vector of row indices in \code{x} identifying the
reference (query) points.}

\item{donors}{An integer vector of row indices in \code{x} identifying the
candidate points.}

\item{n_weights}{An integer in \code{1:length(donors)} defining the number of
donors to be weighted per recipient. Defaults to \code{length(donors)} which
utilises the entire donor pool.}
}
\value{
A numeric vector of length \code{n_weights}.
}
\description{
Compute a single vector of weights which, when applied uniformly for all
\code{recipients}, minimises the aggregate Euclidean distance between imputed and
actual covariates.
}
