% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{derive_local_debiasing_weights}
\alias{derive_local_debiasing_weights}
\title{Derive nearest neighbours weights that reproduce the recipient covariates}
\usage{
derive_local_debiasing_weights(
  x,
  recipients,
  donors,
  n_weights = length(donors)
)
}
\arguments{
\item{x}{A numeric matrix of rowwise covariates used as matching variables
for distance calculations.}

\item{recipients}{An integer vector of row indices in \code{x} identifying the
reference (query) points.}

\item{donors}{An integer vector of row indices in \code{x} identifying the
candidate points.}

\item{n_weights}{An integer in \code{1:length(donors)} defining the number of
donors to be weighted per recipient. Defaults to \code{length(donors)} which
utilises the entire donor pool.}
}
\value{
A \code{length(recipients)}-by-\code{n_weights} numeric matrix of weights
corresponding rowwise to \code{recipients} and columnwise to proximity ordering
\code{1:n_weights}.
}
\description{
Identify vectors of weights for each recipient which, when applied to
weighted nearest neighbours, reproduces the original recipient covariate.
}
