% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kth_nearest.R
\name{fetch_kth_nearest_location}
\alias{fetch_kth_nearest_location}
\title{Extract proximally ordered locations}
\usage{
fetch_kth_nearest_location(recipients, locations, donors, p = 2, k)
}
\arguments{
\item{recipients}{Index vector (or integer) that identifies the recipients
whose responses are to be imputed.}

\item{locations}{Numeric matrix identifying rowwise location of indexes.}

\item{donors}{Index vector of donors. If \code{donors = NULL}, then the donors are
identified by the indexes of non-\code{NA} values in \code{responses}.}

\item{p}{Number >= 1 (or \code{Inf}) specifying which \code{p}-norm to use for
distance calculations. The default is \code{p = 2} (Euclidean distances).}

\item{k}{Positive integer indicating if only the \code{k} nearest donors should
be returned. If \code{k = NULL} (default), then all \code{donors} are ranked.}
}
\value{
\itemize{
\item \code{fetch_kth_nearest_donor_location()} : A row-element of \code{locations} whose
index is in \code{donors}.
\item \code{fetch_all_kth_nearest_donor_locations()}: A matrix with
\code{length(all_recipients)} rows, where each row maps to a row in \code{locations}
whose indexes correspond to elements of \code{donors}.
}
}
\description{
Find the k-th nearest location in a set to a specific point.
}
\details{
The function extracts the location of the \code{k}-th nearest donor for each
\code{recipient}. This is functionally equivalent to \code{\link[=impute_recipient_location]{impute_recipient_location()}}
with \code{weights = c(rep(0, k), 1)}.
}
\seealso{
\code{\link[=order_donors]{order_donors()}} for how distances are calculated.

\code{\link[=impute_location]{impute_location()}} for functional equivalent.
}
