% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{fit_bias_corrected_pseudocensus_ols}
\alias{fit_bias_corrected_pseudocensus_ols}
\title{Compute the bias corrected pseudocensus OLS estimator}
\usage{
fit_bias_corrected_pseudocensus_ols(
  covariates,
  responses,
  donors,
  recipients,
  weights
)
}
\arguments{
\item{covariates}{A rowwise numeric matrix of points.}

\item{responses}{A numeric vector of responses indexed consistently with the
rows of \code{covariates}.}

\item{donors}{An integer vector of row indices in \code{covariates} identifying
the candidate points.  \code{donors} must be disjoint to \code{recipients}.}

\item{recipients}{An integer vector of row indices in \code{covariates}
identifying the reference points. \code{recipients} must be disjoint to \code{donors}.}

\item{weights}{A numeric matrix of imputation weights used for weighted
nearest neighbours imputation. \code{weights} corresponds rowwise to \code{recipients}
and columnwise to the rank of the proximally sorted donors. The number of
weights cannot exceed the number of donors, i.e.,
\code{ncol(weights) <= length(donors)}.}
}
\value{
A numeric vector of length \code{ncol(covariates) + 1}
}
\description{
Compute the bias corrected pseudocensus OLS estimator
}
\details{
Apply the bias correction matrix to the pseudocensus OLS estimator, removing
the bias induced by imputing the responses of \code{recipients} with weighted
nearest neighbours.
}
\seealso{
\code{\link[=fit_pseudocensus_ols]{fit_pseudocensus_ols()}} for the uncorrected estimator and
\code{\link[=bias_correction_matrix]{bias_correction_matrix()}} for the correction applied here.
}
