% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{fit_sample_ols}
\alias{fit_sample_ols}
\title{Compute the ordinary least squares estimator for a population subset}
\usage{
fit_sample_ols(covariates, responses, donors)
}
\arguments{
\item{covariates}{A numeric matrix of rowwise covariates.}

\item{responses}{A numeric vector of length \code{nrow(covariates)}.}

\item{donors}{An integer vector in the row indices of \code{covariates} that
identifies the population subset.}
}
\value{
A numeric vector of length \code{ncol(covariates) + 1} that is the OLS estimate
between \code{covariates} and \code{responses}.
}
\description{
Compute the ordinary least squares estimator for a population subset
}
