% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\alias{impute_covariates}
\alias{impute_responses}
\title{Impute covariates or responses using weighted nearest neighbours}
\usage{
impute_covariates(x, recipients, donors, weights)

impute_responses(x, recipients, donors, weights, y)
}
\arguments{
\item{x}{A numeric matrix of rowwise covariates used as matching variables
for distance calculations.}

\item{recipients}{An integer vector of row indices in \code{x} identifying the
reference (query) points.}

\item{donors}{An integer vector of row indices in \code{x} identifying the
candidate points.}

\item{weights}{A numeric matrix where \code{nrow(weights) = length(recipients)}
and \code{ncol(weights) <= length(donors)}. Each row \code{weights} supplies the
imputation weights for the corresponding recipient. \code{weights} corresponds
columnwise to the rank of the proximally sorted donors.}

\item{y}{A length \code{nrow(x)} numeric vector of responses indexed consistently
with the rows of \code{x}. Only used by \code{impute_responses()}.}
}
\value{
Writing \code{r} = \code{length(recipients)}, \code{p} = \code{ncol(x)}, and \code{k} = \code{ncol(weights)}:
\itemize{
\item \code{impute_covariates()}: an \code{r}-by-\code{p} numeric matrix where each row is the
weighted combination of the \code{k} nearest donor covariates corresponding to
\code{recipients}.
\item \code{impute_responses()}: a numeric vector of length \code{r} where each element is
the weighted combination of responses for the \code{k} nearest donors
corresponding to \code{recipients}.
}
}
\description{
Respectively construct synthetic covariates or responses for a set of
\code{recipients} from weighted combinations of their nearest donor covariates or
responses.
}
\seealso{
\code{\link[=order_donor_indices]{order_donor_indices()}} to see how donors are ranked;
\code{\link[=derive_local_debiasing_weights]{derive_local_debiasing_weights()}} and \code{\link[=derive_global_debiasing_weights]{derive_global_debiasing_weights()}}
for evaluating bias minimising weights.
}
