% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_all_recipient_responses}
\alias{impute_all_recipient_responses}
\alias{impute_recipient_location}
\alias{impute_all_recipient_locations}
\title{Impute a location or response based on weighted k nearest neighbours}
\usage{
impute_all_recipient_responses(
  all_recipients,
  locations,
  responses,
  weights,
  donors = NULL,
  p = 2
)

impute_recipient_location(recipient, donors, locations, weights, p = 2)

impute_all_recipient_locations(
  all_recipients,
  donors,
  locations,
  weights,
  p = 2
)
}
\arguments{
\item{all_recipients}{Index vector that identifies the recipients whose
responses are to be imputed.}

\item{locations}{Numeric matrix identifying rowwise location of indexes.}

\item{responses}{Numeric vector of responses, potentially containing
missing \code{NA} values.}

\item{weights}{Numeric vector of weights used for calculating the weighted
average response or location.}

\item{donors}{Index vector of donors. If \code{donors = NULL}, then the donors are
identified by the indexes of non-\code{NA} values in \code{responses}.}

\item{p}{Number >= 1 (or \code{Inf}) specifying which \code{p}-norm to use for
distance calculations. The default is \code{p = 2} (Euclidean distances).}

\item{recipient}{Index of the recipient.}
}
\value{
\itemize{
\item \code{impute_all_recipient_responses()}: A length \code{length(all_recipients)}
numeric vector of the weighted averages of the \code{length(weights)} nearest
donors for each recipient.
\item \code{impute_recipient_location()}: A numeric vector that approximates the
location of the recipient using the weighted average of the location of the
\code{length(weights)} nearest donors. The length of the output is the
dimension of the location—\code{ncol(locations)}.
\item \code{impute_all_recipient_locations()}: A
\verb{length(all_recipients) × ncol(locations)} matrix where each row contains
the approximated location for a recipient.
}
}
\description{
Supply synthetic approximations for the locations or responses of a
population subset using weighted \eqn{k} nearest neighbours.
}
\details{
\code{impute_all_recipient_responses()} generates responses for each element in
\code{all_recipients} using a weighted average of proximal donor responses.
\code{impute_all_recipient_location()} provides an analogue for the location of
\code{all_recipients}.
}
\seealso{
\code{\link[=order_donors]{order_donors()}} for how distances are calculated.
}
