% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_covariates}
\alias{impute_covariates}
\alias{impute_responses}
\title{Impute covariates or responses using weighted nearest neighbours}
\usage{
impute_covariates(covariates, recipients, donors, weights)

impute_responses(covariates, recipients, donors, weights, responses)
}
\arguments{
\item{covariates}{A rowwise numeric matrix of points.}

\item{recipients}{An integer vector of row indices in \code{covariates}
identifying the reference points. \code{recipients} must be disjoint to \code{donors}.}

\item{donors}{An integer vector of row indices in \code{covariates} identifying
the candidate points.  \code{donors} must be disjoint to \code{recipients}.}

\item{weights}{A numeric matrix of imputation weights used for weighted
nearest neighbours imputation. \code{weights} corresponds rowwise to \code{recipients}
and columnwise to the rank of the proximally sorted donors. The number of
weights cannot exceed the number of donors, i.e.,
\code{ncol(weights) <= length(donors)}.}

\item{responses}{A numeric vector of responses indexed consistently with the
rows of \code{covariates}.}
}
\value{
Writing \code{n_recipients = length(recipients)} and \code{p = ncol(covariates)}:
\itemize{
\item \code{impute_covariates()}: an \code{n_recipients}-by-\code{p} numeric matrix, where each
row is the the weighted combination of nearest donor covariates
corresponding to \code{recipients}.
\item \code{impute_responses()}: a numeric vector of length \code{n_recipients}, where each
element is the weighted combination of the nearest donor responses
corresponding to \code{recipients}.
}
}
\description{
Construct synthetic covariates or responses for a set of recipients from
weighted combinations of their nearest donor values.
}
\seealso{
\code{\link[=order_donor_indices]{order_donor_indices()}} for how donors are ranked;
\code{\link[=derive_local_debiasing_weights]{derive_local_debiasing_weights()}} and \code{\link[=derive_global_debiasing_weights]{derive_global_debiasing_weights()}}
for evaluating bias minimising weights.
}
