% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_response}
\alias{impute_response}
\alias{impute_covariate}
\title{Impute a response based on weighted k nearest neighbours}
\usage{
impute_response(
  recipient,
  covariates,
  responses,
  weights,
  donors = NULL,
  p = 2
)

impute_covariate(recipient, donors, covariates, weights, p = 2)
}
\arguments{
\item{recipient}{Integer index identifying the recipient in \code{covariates}.}

\item{covariates}{Numeric matrix where each row is an individual's covariates.}

\item{responses}{Numeric vector of responses, pontentitally containing
missing \code{NA} values}

\item{weights}{Numeric vector of weights used for calculating the weighted
average.}

\item{donors}{Integer vector of indices identifying the donors in \code{covariates}.}

\item{p}{A double >= 1 (or \code{Inf}) specifying which \code{p}-norm to use for
distance calculations. The deafult is \code{p = 2} (Euclidean distances).}
}
\value{
\itemize{
\item \code{impute_response()}: A numeric value of the weighted average of the
responses of the \code{length(weights)} nearest donors.
\item \code{impute_covariate()}: A numeric vector of the weighted average of the
covariates of the \code{length(weights)} nearest donors.
}
}
\description{
\code{impute_response()} constructs a synthetic response for the \code{recipient} by
calculating a weighted average of proximal \code{donor} responses—using \eqn{k}
nearest neighbours. \code{impute_covariate()} supplies the equivalent for the
covariate vector of the \code{recipient}.
}
