% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge}
\alias{merge}
\alias{merge_covariates}
\alias{merge_responses}
\title{Impute recipient values in place by weighted nearest neighbours}
\usage{
merge_covariates(x, recipients, donors, weights)

merge_responses(x, recipients, donors, weights, y)
}
\arguments{
\item{x}{A numeric matrix of rowwise covariates used as matching variables
for distance calculations.}

\item{recipients}{An integer vector of row indices in \code{x} identifying the
reference (query) points.}

\item{donors}{An integer vector of row indices in \code{x} identifying the
candidate points.}

\item{weights}{A numeric matrix where \code{nrow(weights) = length(recipients)}
and \code{ncol(weights) <= length(donors)}. Each row \code{weights} supplies the
imputation weights for the corresponding recipient. \code{weights} corresponds
columnwise to the rank of the proximally sorted donors.}

\item{y}{A length \code{nrow(x)} numeric vector of responses indexed consistently
with the rows of \code{x}. Only used by \code{impute_responses()}.}
}
\value{
\itemize{
\item \code{merge_covariates()}: A numeric matrix with the same dimensions as
\code{x}, in which rows indexed by \code{recipients} are replaced with their weighted
nearest neighbours estimate and all other rows are unchanged.
\item \code{merge_responses()}: A numeric vector with the same length as \code{y},
where elements indexed by \code{recipients} are replaced with their weighted
nearest neighbours estimate and all other elements are unchanged.
}
}
\description{
Impute the covariates or responses of \code{recipients} from weighted combinations
of their proximally ranked donors covariates or responses, and splice the
estimates back into the full data object in place of the original recipient
entries.
}
\seealso{
\code{\link[=order_donor_indices]{order_donor_indices()}} to see how donors are ranked;
\code{\link[=derive_local_debiasing_weights]{derive_local_debiasing_weights()}} and \code{\link[=derive_global_debiasing_weights]{derive_global_debiasing_weights()}}
for constructing bias minimising weights.
}
