% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_covariates}
\alias{merge_covariates}
\alias{merge_responses}
\title{Replace only recipient values using weighted nearest neighbours}
\usage{
merge_covariates(covariates, recipients, donors, weights)

merge_responses(covariates, recipients, donors, weights, responses)
}
\arguments{
\item{covariates}{A rowwise numeric matrix of points.}

\item{recipients}{An integer vector of row indices in \code{covariates}
identifying the reference points. \code{recipients} must be disjoint to \code{donors}.}

\item{donors}{An integer vector of row indices in \code{covariates} identifying
the candidate points.  \code{donors} must be disjoint to \code{recipients}.}

\item{weights}{A numeric matrix of imputation weights used for weighted
nearest neighbours imputation. \code{weights} corresponds rowwise to \code{recipients}
and columnwise to the rank of the proximally sorted donors. The number of
weights cannot exceed the number of donors, i.e.,
\code{ncol(weights) <= length(donors)}.}

\item{responses}{A numeric vector of responses indexed consistently with the
rows of \code{covariates}.}
}
\value{
\itemize{
\item \code{merge_covariates()}: A numeric matrix of the same dimension as
\code{covariates}, where only rows indexed by \code{recipients} are replaced with
their nearest neighbours estimate. Rows indexed by \code{donors} are unchanged.
\item \code{merge_responses()}: A numeric vector of the same length as \code{responses}
where only elements indexed by \code{recipients} are replaced with their nearest
neighbours estimate. Rows indexed by \code{donors} remain unchanged.
}
}
\description{
Replace only recipient values using weighted nearest neighbours
}
