% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_response}
\alias{merge_response}
\title{Merge imputed and observed responses and locations}
\usage{
merge_response(
  locations,
  responses,
  weights,
  p = 2,
  all_recipients = NULL,
  donors = NULL
)
}
\arguments{
\item{locations}{Numeric matrix identifying rowwise location of indexes.}

\item{responses}{Numeric vector of responses, potentially containing
missing \code{NA} values.}

\item{weights}{Numeric vector of weights used for calculating the weighted
average response or location.}

\item{p}{Number >= 1 (or \code{Inf}) specifying which \code{p}-norm to use for
distance calculations. The default is \code{p = 2} (Euclidean distances).}

\item{donors}{Index vector of donors. If \code{donors = NULL}, then the donors are
identified by the indexes of non-\code{NA} values in \code{responses}.}
}
\value{
\itemize{
\item \code{merge_response()}: returns \code{responses}, except where the responses for
the recipient indices have been replaced with their imputed equivalent.
\item \code{merge_location()}: returns \code{locations}, except where the rows for the
recipient indices have been replaced with their imputed equivalent.
}
}
\description{
Combine observed and imputed responses (or locations) into a single array.
}
