% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R
\name{order_donor_indices}
\alias{order_donor_indices}
\alias{order_donor_covariates}
\alias{order_donor_responses}
\title{Order donors by their distance from recipients}
\usage{
order_donor_indices(covariates, recipients, donors, k = NULL)

order_donor_covariates(covariates, recipients, donors, k = NULL)

order_donor_responses(covariates, recipients, donors, responses, k = NULL)
}
\arguments{
\item{covariates}{A rowwise numeric matrix of points.}

\item{recipients}{An integer vector of row indices in \code{covariates}
identifying the reference points. \code{recipients} must be disjoint to \code{donors}.}

\item{donors}{An integer vector of row indices in \code{covariates} identifying
the candidate points.  \code{donors} must be disjoint to \code{recipients}.}

\item{k}{An integer scalar between 1 and \code{length(donors)} giving the number
of donors to return per recipient. The default sets \code{k = length(donors)}
and rank the full list of donors for each recipient.}

\item{responses}{A numeric vector of responses indexed consistently with the
rows of \code{covariates}.}
}
\value{
Writing \code{n_recipients = length(recipients)} and \code{p = ncol(covariates)}:
\itemize{
\item \code{order_donor_indices()}: an \code{n_recipients}-by-\code{k} integer matrix where
each row is a permuted subset of \code{donors} where, for the corresponding
recipient index, the permutation sorts the donors by their proximity from
the recipient.
\item \code{order_donor_covariates()}: a length \code{n_recipients} list of \code{k}-by-\code{p}
numeric matrices. Each element in the list is a rowwise permuted subset of
\code{covariates} corresponding to the index permutation in
\code{order_donor_indices()}.
\item \code{order_donor_responses()}: an \code{n_recipients}-by-\code{k} numeric matrix where
each row is a permuted subset of \code{responses} corresponding to the index
permutation in \code{order_donor_indices()}.
}
}
\description{
Rank a candidate set of points by their distance from a set of reference
points and return their associated indices, covariates, or responses.
}
\details{
\verb{order_donor_*()} defines a family of functions that ranks a candidate set
of points by their distance to a set of reference points. The points are
the \code{ncol(covariates)}-dimensional vectors defined by the rows of
\code{covariates}. The disjoint index vectors \code{recipients} and \code{donors}
respectively define the indexes of the reference and candidate points. Donors
are independently ranked by proximity for each recipient, with respect to the
Euclidean metric via \code{\link[nabor:knn]{nabor::knn()}}. The associated ordered set of indices,
covariates, or responses are returned depending on which of \verb{order_donor_*()}
is called.
}
\seealso{
\code{\link[nabor:knn]{nabor::knn()}} for the underlying nearest-neighbour search.
}
