% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_correction_matrix.R
\name{bias_correction_matrix}
\alias{bias_correction_matrix}
\title{Evaluate the bias correction matrix for the pseudocensus OLS estimator}
\usage{
bias_correction_matrix(covariates, recipients, donors, weights)
}
\arguments{
\item{covariates}{A rowwise numeric matrix of points.}

\item{recipients}{An integer vector of row indices in \code{covariates}
identifying the reference points. \code{recipients} must be disjoint to \code{donors}.}

\item{donors}{An integer vector of row indices in \code{covariates} identifying
the candidate points.  \code{donors} must be disjoint to \code{recipients}.}

\item{weights}{A numeric matrix of imputation weights used for weighted
nearest neighbours imputation. \code{weights} corresponds rowwise to \code{recipients}
and columnwise to the rank of the proximally sorted donors. The number of
weights cannot exceed the number of donors, i.e.,
\code{ncol(weights) <= length(donors)}.}
}
\value{
A square numeric matrix with \code{ncol(covariates)} rows and columns.
}
\description{
Evaluate the bias correction matrix for the pseudocensus OLS estimator
}
