% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{derive_local_debiasing_weights}
\alias{derive_local_debiasing_weights}
\title{Derive weights that recover the recipient covariates}
\usage{
derive_local_debiasing_weights(
  covariates,
  recipients,
  donors,
  n_weights = NULL
)
}
\arguments{
\item{covariates}{A rowwise numeric matrix of points.}

\item{recipients}{An integer vector of row indices in \code{covariates}
identifying the reference points. \code{recipients} must be disjoint to \code{donors}.}

\item{donors}{An integer vector of row indices in \code{covariates} identifying
the candidate points.  \code{donors} must be disjoint to \code{recipients}.}

\item{n_weights}{A positive integer defining the number of weights used in
in the weighted nearest neighbours estimates, uniformly for all \code{recipients}.
The number of weights cannot exceed the number of valid donors, i.e.,
\code{n_weights <= length(donors)}. The default argument sets
\code{n_weights = length(donors)}.}
}
\value{
A \code{length(recipients)}-by-\code{n_weights} numeric matrix of weights
corresponding rowwise to \code{recipients} and columnwise to proximity ordering
\code{1:n_weights}.
}
\description{
Identify vectors of weights for each recipient which, when applied to
weighted nearest neighbours estimates, recovers the recipient covariate.
}
