% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{fit_pseudocensus_ols}
\alias{fit_pseudocensus_ols}
\title{Compute the ordinary least squares estimator for a partially nearest
neighbours imputed population}
\usage{
fit_pseudocensus_ols(x, y, donors, recipients, weights)
}
\arguments{
\item{x}{A numeric matrix of rowwise covariates used as matching variables
for distance calculations.}

\item{y}{A length \code{nrow(x)} numeric vector of responses indexed consistently
with the rows of \code{x}. Only used by \code{impute_responses()}.}

\item{donors}{An integer vector of row indices in \code{x} identifying the
candidate points.}

\item{recipients}{An integer vector of row indices in \code{x} identifying the
reference (query) points.}

\item{weights}{A numeric matrix where \code{nrow(weights) = length(recipients)}
and \code{ncol(weights) <= length(donors)}. Each row \code{weights} supplies the
imputation weights for the corresponding recipient. \code{weights} corresponds
columnwise to the rank of the proximally sorted donors.}
}
\value{
A numeric vector of length \code{ncol(x)}.
}
\description{
Compute the ordinary least squares estimator for a partially nearest
neighbours imputed population
}
\details{
Compute the ordinary least squares estimator in which a population subset of
\code{recipients} has had their responses imputed by weighted nearest neighbours.
}
