% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{fit_pseudocensus_ols}
\alias{fit_pseudocensus_ols}
\title{Compute the ordinary least squares estimator for a partially nearest
neighbours imputed population}
\usage{
fit_pseudocensus_ols(covariates, responses, donors, recipients, weights)
}
\arguments{
\item{covariates}{A rowwise numeric matrix of points.}

\item{responses}{A numeric vector of responses indexed consistently with the
rows of \code{covariates}.}

\item{donors}{An integer vector of row indices in \code{covariates} identifying
the candidate points.  \code{donors} must be disjoint to \code{recipients}.}

\item{recipients}{An integer vector of row indices in \code{covariates}
identifying the reference points. \code{recipients} must be disjoint to \code{donors}.}

\item{weights}{A numeric matrix of imputation weights used for weighted
nearest neighbours imputation. \code{weights} corresponds rowwise to \code{recipients}
and columnwise to the rank of the proximally sorted donors. The number of
weights cannot exceed the number of donors, i.e.,
\code{ncol(weights) <= length(donors)}.}
}
\value{
A numeric vector of length \code{ncol(covariates) + 1}
}
\description{
Compute the ordinary least squares estimator for a partially nearest
neighbours imputed population
}
\details{
Compute the ordinary least squares estimator in which a population subset of
\code{recipients} has had their responses imputed by weighted nearest neighbours
}
