% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{fit_sample_ols}
\alias{fit_sample_ols}
\title{Compute the OLS estimator for a population subset}
\usage{
fit_sample_ols(x, y, donors)
}
\arguments{
\item{x}{A numeric matrix of rowwise predictors.}

\item{y}{A numeric vector of responses of length \code{nrow(x)}.}

\item{donors}{An integer vector in the row indices of \code{x} that identifies the
population subset.}
}
\value{
A numeric vector of length \code{ncol(x)}
}
\description{
Compute the OLS estimator for a population subset
}
