% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify_donors}
\alias{identify_donors}
\alias{identify_recipients}
\title{Identify donors and recipients in a population}
\usage{
identify_donors(responses)

identify_recipients(responses)
}
\arguments{
\item{responses}{A numeric vector that may contain missing elements (\code{NA} values).}
}
\value{
\itemize{
\item \code{identify_donors()}: Integer vector of indices where responses are not \code{NA}.
\item \code{identify_recipients()}: Integer vector of indices where responses are \code{NA}.
}
}
\description{
\code{identify_donors()} and \code{identify_recipients()} respectively identify the
respondent (donors) and nonrespondent (recipient) subsets of a population
based on missing responses.
}
