% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_response}
\alias{impute_response}
\alias{impute_location}
\title{Impute a location or response based on weighted k nearest neighbours}
\usage{
impute_response(
  recipients,
  locations,
  responses,
  weights,
  donors = NULL,
  p = 2
)

impute_location(recipients, donors, locations, weights, p = 2)
}
\arguments{
\item{recipients}{Index vector (or integer) that identifies the recipients
whose responses are to be imputed.}

\item{locations}{Numeric matrix identifying rowwise location of indexes.}

\item{responses}{Numeric vector of responses, potentially containing
missing \code{NA} values.}

\item{weights}{Numeric vector of weights used for calculating the weighted
average response or location.}

\item{donors}{Index vector of donors. If \code{donors = NULL}, then the donors are
identified by the indexes of non-\code{NA} values in \code{responses}.}

\item{p}{Number >= 1 (or \code{Inf}) specifying which \code{p}-norm to use for
distance calculations. The default is \code{p = 2} (Euclidean distances).}
}
\value{
\itemize{
\item \code{impute_response()}: A length \code{length(recipients)} numeric vector of the
weighted averages of the \code{length(weights)} nearest donors for each
recipient.
\item \code{impute_location()}: A \verb{length(recipients) × ncol(locations)} matrix where
each row contains the approximated location for a recipient.
}
}
\description{
Supply synthetic approximations for the locations or responses of a
population subset using weighted \eqn{k} nearest neighbours.
}
\details{
\code{impute_response()} generates responses for each element in \code{recipients}
using a weighted average of proximal donor responses. \code{impute_location()}
provides an analogue for the location of \code{recipients}.
}
\seealso{
\code{\link[=order_donors]{order_donors()}} for how distances are calculated.
}
