% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_responses}
\alias{merge_responses}
\alias{merge_covariates}
\title{Merge imputed and observed responses and covariates}
\usage{
merge_responses(
  covariates,
  responses,
  weights,
  p = 2,
  recipients = NULL,
  donors = NULL
)

merge_covariates(
  covariates,
  weights,
  p = 2,
  recipients = NULL,
  donors = NULL,
  responses = NULL
)
}
\arguments{
\item{covariates}{Numeric matrix where each row is an individual's covariates.}

\item{responses}{Numeric vector of responses, pontentitally containing
missing \code{NA} values}

\item{weights}{Numeric vector of weights used for calculating the weighted
average.}

\item{p}{A double >= 1 (or \code{Inf}) specifying which \code{p}-norm to use for
distance calculations. The deafult is \code{p = 2} (Euclidean distances).}

\item{recipients}{Integer vector of recipients whose responses, possibly
missing, require imputation}

\item{donors}{Integer vector of indices identifying the donors in \code{covariates}.}
}
\value{
\itemize{
\item \code{merge_responses()}: returns \code{responses}, except where the responses for
the recipient indices have been replaced with their weighted \eqn{k}
nearest neighbours imputed equivalent.
\item \code{merge_covariates()}: returns \code{covariates}, except where the rows for the
recipient indices have been replaced with their weighted \eqn{k}
nearest neighbours imputed equivalent.
}
}
\description{
Combine observed and imputed responses (or covariates) into a single object.
}
