% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_donors.R
\name{order_donors}
\alias{order_donors}
\title{Rank the donors by their distance from a recipient}
\usage{
order_donors(recipient, donors, covariates, p = 2)
}
\arguments{
\item{recipient}{Integer index identifying the recipient in \code{covariates}.}

\item{donors}{Integer vector of indices identifying the donors in \code{covariates}.}

\item{covariates}{Numeric matrix where each row is an individual's covariates.}

\item{p}{A double >= 1 (or \code{Inf}) specifying which \code{p}-norm to use for
distance calculations. The deafult is \code{p = 2} (Euclidean distances).}
}
\value{
A permutation of the integer donor indices according to their
ascendingly ranked proximity to the recipient.
}
\description{
Order donors according to their proximity to a recipient using the \code{p}-norm
distance between their respective covariates.
}
\details{
\code{order_donors()} ascendingly ranks the donors according to their proximity to
a recipient based on their covariates with respect to the vector \code{p}-norm.
}
